@@ -14,27 +14,31 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.0.6 <4.0.0"
 
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  font_awesome_flutter: 
+  cupertino_icons: ^1.0.2
+  font_awesome_flutter:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -57,25 +61,25 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
-  - family: Raleway
-    fonts:
-      - asset: assets/fonts/Raleway/Raleway-Regular.ttf
-      - asset: assets/fonts/Raleway/Raleway-Light.ttf
-        weight: 100
-      - asset: assets/fonts/Raleway/Raleway-Bold.ttf
-        weight: 700
-  - family: Montserrat
-    fonts:
-      - asset: assets/fonts/Montserrat/Montserrat-Regular.ttf
-      - asset: assets/fonts/Montserrat/Montserrat-Light.ttf
-        weight: 100
-      - asset: assets/fonts/Montserrat/Montserrat-Bold.ttf
-        weight: 700
+    - family: Raleway
+      fonts:
+        - asset: assets/fonts/Raleway/Raleway-Regular.ttf
+        - asset: assets/fonts/Raleway/Raleway-Light.ttf
+          weight: 100
+        - asset: assets/fonts/Raleway/Raleway-Bold.ttf
+          weight: 700
+    - family: Montserrat
+      fonts:
+        - asset: assets/fonts/Montserrat/Montserrat-Regular.ttf
+        - asset: assets/fonts/Montserrat/Montserrat-Light.ttf
+          weight: 100
+        - asset: assets/fonts/Montserrat/Montserrat-Bold.ttf
+          weight: 700
 
-  - family: Lato
-    fonts:
-      - asset: assets/fonts/Lato/Lato-Regular.ttf
-      - asset: assets/fonts/Lato/Lato-Light.ttf
-        weight: 100
-      - asset: assets/fonts/Lato/Lato-Bold.ttf
-        weight: 700
+    - family: Lato
+      fonts:
+        - asset: assets/fonts/Lato/Lato-Regular.ttf
+        - asset: assets/fonts/Lato/Lato-Light.ttf
+          weight: 100
+        - asset: assets/fonts/Lato/Lato-Bold.ttf
+          weight: 700
